Added permissions gate to deleteItem resolver
deleteItem now checks whether user has deleting rights

+++ src/resolvers/Mutation.js
@@ -42,11 +42,18 @@ const mutations = {
     },
 
     async deleteItem(parent, args, ctx, info) {
-        const where = {id : args.id};
         //1. find the item
-        const item = await ctx.db.query.item({ where }, `{id title}`);
+        const where = { id : args.id };
+        const item = await ctx.db.query.item({ where }, `{id title user {id} }`);
+    
         //2. Check if they own the item
-        //TODO
+        const ownsItem = item.user.id === ctx.request.userID;
+        const hasPermission = ctx.request.user.permissions.some(permission => ["ITEMDELETE", "ADMIN"].includes(permission));
+
+        if(!ownsItem && !hasPermission) {
+            throw new Error("You don't have permission to do that");
+        }
+
         //3. Delete it
         return ctx.db.mutation.deleteItem({ where }, info);
     },
